requirements.txt
torch==1.11.0
PettingZoo[mpe]==1.18.1
PettingZoo[classic]==1.18.1
numpy==1.22.3
matplotlib==3.5.2
Pillow==9.1.1
gym==0.23.1
pygame
gymnasium
pyglet==1.5